% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_stats.R
\name{calculate_stats}
\alias{calculate_stats}
\title{Calculate record, win percentage, and rank for each team}
\usage{
calculate_stats(schedule, scores, league = c("Yahoo", "ESPN"))
}
\arguments{
\item{schedule}{Dataframe of schedule for entire season.}

\item{league}{type of league.}

\item{score}{Dataframe of weekly scores.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Calculate record, win percentage, and rank for each team
}
